+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.7
 - (Trade Window) Allow adding descriptions to presets
 - (Trade Window) Apply correct item quality when selecting preset
+- (Misc) When crafting sub-spirit gear, the Manual tooltip now contains all the information related to that manual
 
 ## v0.6.2 (06/05/2026)
 - (Misc) Fix loop shortcut logic